bugfix: Fixed NPEs discovered in BugSense
https://www.bugsense.com/dashboard/project/ad73db70#error/97976122
https://www.bugsense.com/dashboard/project/ad73db70#error/97551051
https://www.bugsense.com/dashboard/project/ad73db70#error/97340424
https://www.bugsense.com/dashboard/project/ad73db70#error/97976122
https://www.bugsense.com/dashboard/project/ad73db70#error/97340424

+++ src/paulscode/android/mupen64plusae/util/Utility.java
@@ -168,9 +168,12 @@ public class Utility
             
             // Clear the folder if anything is in there:
             String[] children = tmpFolder.list();
-            for( String child : children )
+            if( children != null )
             {
-                FileUtil.deleteFolder( new File( tmpFolder, child ) );
+                for( String child : children )
+                {
+                    FileUtil.deleteFolder( new File( tmpFolder, child ) );
+                }
             }
             
             ErrorLogger.clearLastError();
@@ -242,9 +245,12 @@ public class Utility
             
             // Clear the folder if anything is in there:
             String[] children = tmpFolder.list();
-            for( String child : children )
+            if( children != null)
             {
-                FileUtil.deleteFolder( new File( tmpFolder, child ) );
+                for( String child : children )
+                {
+                    FileUtil.deleteFolder( new File( tmpFolder, child ) );
+                }
             }
             
             ErrorLogger.clearLastError();
